// Desktop/client-mis/types/types.ts
export interface IProductMainPage {
  id: number;
  brand: {
    id: number;
    image: string;
    name: string;
  };
  color: {
    id: number;
    name: string;
    color: string;
  };
  size: {
    id: number;
    name: string;
  };
  manufacturerCountry: {
    id: number;
    name: string;
  };
  subTypes: {
    id: number;
    name: string;
    type: number;
  }[];
  first_image: string;
  second_image: string;
  third_image: string;
  fourth_image: string;
  fifth_image: string;
  name: string;
  description: string;
  article_number: string;
  amount: number;
  compound: string;
  price: string;
  season: string;
  pattern: string;
  slug: string;
  is_famous: boolean;
  date_and_time: string;
}
// ...
export interface IProductsState {
  products: IProductMainPage[];
  filtered: IProductMainPage[];
  category: string;
  status: IFavoritesState["status"];
  searchText: string;
  filters: {
    size: string | null;
    brand: string | null;
    color: string | null;
    lbprice: number | null;
    hbprice: number | null;
  };
}

export interface IFavoritesState {
  favorites: IProductMainPage[];
  status: "pending" | "fulfilled";
}